Refactor and split the tokens
If we are using an external AI system, then we need a token for AI inference and a token for GitHub interactions.
