+++ __fixtures__/ai.ts
@@ -4,5 +4,4 @@
 import type * as ai from '../src/ai.js'
 import { jest } from '@jest/globals'
 
-export const generatePrompt = jest.fn<typeof ai.generatePrompt>()
 export const runInference = jest.fn<typeof ai.runInference>()
