+++ __fixtures__/prompts-summary.ts
@@ -1,7 +1,7 @@
 /**
- * Mock fixture for summary functions (src/summary.ts)
+ * Mock fixture for summary functions (src/prompts-summary.ts)
  */
-import type * as summary from '../src/summary.js'
+import type * as summary from '../src/prompts-summary.js'
 import { jest } from '@jest/globals'
 
 export const generateSummary = jest.fn<typeof summary.generateSummary>()
