+++ src/prompts-summary.ts
@@ -2,7 +2,8 @@ import * as core from '@actions/core'
 import * as fs from 'fs'
 import * as path from 'path'
 import { getPrompt } from './prompts/apply/index.js'
-import { runInference, generatePrompt } from './ai.js'
+import { generatePrompt } from './prompts.js'
+import { runInference } from './ai.js'
 import { SummaryPromptConfig } from './triage-config.js'
 import { TriageResponse } from './triage-response.js'
